+++ Source/PC/Graphics_PC.cpp
@@ -21,6 +21,7 @@
  */
 
 #include "stdafx.hpp"
+#include "PC/SpriteData_PC.hpp"
 
 cGraphics_PC::~cGraphics_PC() {
 
@@ -662,7 +663,7 @@ void cGraphics_PC::Recruit_Draw_HomeAway( ) {
 	mFodder->sub_13C1C( 0x0E, 0x0A, 0, 0x9B );
 	
 	std::string Home = tool_StripLeadingZero(tool_NumToString( mFodder->word_397AE ));
-	mFodder->sub_16B55( 0x0D, 0x9A - (Home.length() * 0x0C), 0x0A, Home );
+	mFodder->sub_16B55( 0x0D, (int16) 0x9A - (Home.length() * 0x0C), 0x0A, Home );
 
 	std::string Away = tool_StripLeadingZero(tool_NumToString( mFodder->word_397AC ));
 	mFodder->sub_16B55( 0x0D, 0xAA, 0x0A, Away );
